requirements.txt
asgiref==3.9.0
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
factory_boy==3.3.3
Faker==37.4.0
python-decouple==3.8
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
